+++ example/example/lib/stories/zwap_checkbox_story.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:zwap_design_system/atoms/atoms.dart';
 import 'package:zwap_design_system/atoms/checkbox/zwapCheckbox.dart';
 import 'package:zwap_design_system/atoms/text/base/zwapText.dart';
+import 'package:zwap_design_system/molecules/molecules.dart';
 
 class ZwapCheckBoxStory extends StatefulWidget {
   const ZwapCheckBoxStory({Key? key}) : super(key: key);
@@ -69,6 +70,20 @@ class _ZwapCheckBoxStoryState extends State<ZwapCheckBoxStory> {
                 error: _error,
                 onCheckBoxClick: (value) => setState(() => _value = value),
               ),
+              SizedBox(height: 40),
+              ZwapText(
+                text: "CheckBox Text",
+                zwapTextType: ZwapTextType.bigBodyRegular,
+                textColor: ZwapColors.shades100,
+              ),
+              SizedBox(height: 15),
+              //TODO: refactoring of ZwapCheckBoxText
+              /* ZwapCheckBoxText(
+                value: _value,
+                disabled: _disabled,
+                error: _error,
+                onCheckBoxClick: (value) => setState(() => _value = value),
+              ), */
             ],
           ),
         ],