+++ CHANGELOG.md
@@ -573,3 +573,7 @@
 
 ## 0.1.558
 * Created ZwapGradientText and ZwapGradientTextSpan
+
+## 0.1.559
+* Created ZwapHintInput widget to let user enter tags typing
+* Added completionValue logics to ZwapButton